pyproject.toml
[project]
name = "logic-graph"
version = "0.1.0"
description = "Logic graph for a psychoeducation chatbot about depression, Colombian context"
authors = [
    { name = "DanielPantoja08", email = "danny0897pant@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-google-genai>=2",
    "langchain-groq>=0.3",
    "langchain-ollama>=0.3",
    "pydantic>=2",
    "pydantic-settings>=2",
    "python-dotenv>=1",
    "fastapi>=0.115",
    "fastapi-users[sqlalchemy]>=14",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
    "langgraph-checkpoint-postgres>=2",
    "alembic>=1.14",
    "chromadb>=1.5.5",
    "google-genai>=1.68.0",
    "docling>=2.82.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "mcp[cli]>=1.0",
    "langsmith>=0.2",
]

[project.scripts]
logic-graph = "logic_graph.app:main"
logic-graph-api = "logic_graph.api.main:run"
logic-graph-mcp = "logic_graph.mcp_rag.server:main"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: requires a running PostgreSQL instance",
]

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"
